refactor: Split event store maintenance interface
