+++ eventstore/mongodb/eventstore.go
@@ -278,87 +278,6 @@ func (s *EventStore) Load(ctx context.Context, id uuid.UUID) ([]eh.Event, error)
 	return events, nil
 }
 
-// Replace implements the Replace method of the eventhorizon.EventStore interface.
-func (s *EventStore) Replace(ctx context.Context, event eh.Event) error {
-	c := s.client.Database(s.dbName(ctx)).Collection("events")
-
-	// First check if the aggregate exists, the not found error in the update
-	// query can mean both that the aggregate or the event is not found.
-	if n, err := c.CountDocuments(ctx, bson.M{"_id": event.AggregateID()}); n == 0 {
-		return eh.ErrAggregateNotFound
-	} else if err != nil {
-		return eh.EventStoreError{
-			Err:       err,
-			Namespace: eh.NamespaceFromContext(ctx),
-		}
-	}
-
-	// Create the event record for the Database.
-	e, err := newEvt(ctx, event)
-	if err != nil {
-		return err
-	}
-
-	// Find and replace the event.
-	if r, err := c.UpdateOne(ctx,
-		bson.M{
-			"_id":            event.AggregateID(),
-			"events.version": event.Version(),
-		},
-		bson.M{
-			"$set": bson.M{"events.$": *e},
-		},
-	); err != nil {
-		return eh.EventStoreError{
-			Err:       eh.ErrCouldNotSaveEvents,
-			BaseErr:   err,
-			Namespace: eh.NamespaceFromContext(ctx),
-		}
-	} else if r.MatchedCount == 0 {
-		return eh.ErrInvalidEvent
-	}
-
-	return nil
-}
-
-// RenameEvent implements the RenameEvent method of the eventhorizon.EventStore interface.
-func (s *EventStore) RenameEvent(ctx context.Context, from, to eh.EventType) error {
-	c := s.client.Database(s.dbName(ctx)).Collection("events")
-
-	// Find and rename all events.
-	// TODO: Maybe use change info.
-	if _, err := c.UpdateMany(ctx,
-		bson.M{
-			"events.event_type": from.String(),
-		},
-		bson.M{
-			"$set": bson.M{"events.$.event_type": to.String()},
-		},
-	); err != nil {
-		return eh.EventStoreError{
-			Err:       eh.ErrCouldNotSaveEvents,
-			BaseErr:   err,
-			Namespace: eh.NamespaceFromContext(ctx),
-		}
-	}
-
-	return nil
-}
-
-// Clear clears the event storage.
-func (s *EventStore) Clear(ctx context.Context) error {
-	c := s.client.Database(s.dbName(ctx)).Collection("events")
-
-	if err := c.Drop(ctx); err != nil {
-		return eh.EventStoreError{
-			Err:       ErrCouldNotClearDB,
-			BaseErr:   err,
-			Namespace: eh.NamespaceFromContext(ctx),
-		}
-	}
-	return nil
-}
-
 // Close closes the database client.
 func (s *EventStore) Close(ctx context.Context) {
 	s.client.Disconnect(ctx)
